requirements.txt
jupyter
scikit-learn
pandas
numpy
matplotlib
seaborn
torch
beautifulsoup4
tqdm
spacy
scipy
nltk
gensim
tweet-preprocessor
pyaml
pytorch_pretrained_bert
torchtext==0.4.0
vaderSentiment
praw
psaw
transformers
pytorch_transformers